Decorator proposal for individual storages - low level storages (binary, directory (filesystem)) - resource units storages (files, rpms, tarballs) - decorated storages (rpmstorage, specfilestorage, githubsourcecodestorage, etc.)
When storing upstream source code I don't want to have my class to decide if the source code
is going to be stored in a file system directory or in binary form in a database.

I would like to let user to have the responsibility for deciding the form of storing the resources
(as a file in a filesystem or as binary data in database).

This way, individual storages are more reusable components which can be combined to get the desired functionality.
For each use case I will have a recipe (see example.py)